requirements.txt
# Core dependencies
pydantic>=2.0.0
google-generativeai>=0.3.0

# Web interface
flask>=3.0.0

# Development (optional)
pytest>=7.4.0
